refactor: move runtime_assets.py from environment to utils
Fix layer architecture violation where adapters (L6) imported from
environment (L5), causing upward dependency violation.

Changes:
- Move runtime_assets.py from environment/ to utils/ (L6)
- Update all import paths across codebase
- Add re-exports in environment/__init__.py for backward compatibility
- Update test mocks to use new module path

This resolves CI failure in test_no_upward_imports.

Layer architecture:
- Layer 6 (utils): Infrastructure - can be imported by all layers
- Layer 5 (environment): Environment tools - cannot be imported by L6
