+++ src/vibe3/environment/__init__.py
@@ -1,10 +1,5 @@
 """Environment isolation modules (worktree, session, runtime assets)."""
 
-from vibe3.environment.runtime_assets import (
-    resolve_prompt_config,
-    resolve_runtime_asset,
-    runtime_assets_root,
-)
 from vibe3.environment.session import (
     CodeagentSessionContext,
     SessionManager,
@@ -13,6 +8,11 @@ from vibe3.environment.session import (
 from vibe3.environment.session_registry import SessionRegistryService
 from vibe3.environment.worktree import WorktreeManager
 from vibe3.environment.worktree_context import WorktreeContext
+from vibe3.utils.runtime_assets import (
+    resolve_prompt_config,
+    resolve_runtime_asset,
+    runtime_assets_root,
+)
 
 __all__ = [
     "WorktreeContext",
