+++ src/vibe3/roles/run_command.py
@@ -17,7 +17,6 @@ from vibe3.clients.sqlite_client import SQLiteClient
 from vibe3.config.convention_resolver import ConventionResolver
 from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.config.settings import VibeConfig
-from vibe3.environment.runtime_assets import resolve_runtime_asset
 from vibe3.exceptions import SkillNotAvailableError
 from vibe3.execution.codeagent_runner import CodeagentExecutionService
 from vibe3.execution.codeagent_support import build_self_invocation
@@ -32,6 +31,7 @@ from vibe3.roles.run_helpers import (
     publish_run_command_success,
 )
 from vibe3.services.error_helpers import record_dispatch_failure_if_unexpected
+from vibe3.utils.runtime_assets import resolve_runtime_asset
 
 
 def resolve_skill_path(
