+++ src/vibe3/utils/__init__.py
@@ -5,9 +5,17 @@ from vibe3.utils.constants import (
     GENERIC_AGENT_MARKER_PATTERN,
     STARTING_TIMEOUT_SECONDS,
 )
+from vibe3.utils.runtime_assets import (
+    resolve_prompt_config,
+    resolve_runtime_asset,
+    runtime_assets_root,
+)
 
 __all__ = [
     "AUTOMATED_MARKERS",
     "GENERIC_AGENT_MARKER_PATTERN",
     "STARTING_TIMEOUT_SECONDS",
+    "resolve_prompt_config",
+    "resolve_runtime_asset",
+    "runtime_assets_root",
 ]
