+++ src/vibe3/utils/runtime_assets.py
