+++ tests/vibe3/integration/test_install_runtime_assets.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-from vibe3.environment.runtime_assets import runtime_assets_root
+from vibe3.utils.runtime_assets import runtime_assets_root
 
 
 def test_runtime_assets_root_points_to_vibe_home(monkeypatch, tmp_path: Path) -> None:
